added common funcs to ppocr examples.  added benchmark to ppocr detect example
